Revert "Update keyboards/keyboardio/atreus/keymaps/kkokdae/keymap.c"
This reverts commit 8e9ccc78c796ea482b5b48d546498c06e3a2af09.

+++ keyboards/keyboardio/atreus/keymaps/kkokdae/keymap.c
@@ -104,7 +104,7 @@ void dance_hex(qk_tap_dance_state_t *state, void *user_data) {
 }
 
 void dance_lang(qk_tap_dance_state_t *state, void *user_data) {
-    uint8_t default_layer;
+  uint32_t default_layer;
   switch(state->count) {
     case 1:
 			SEND_STRING(SS_LCTL(" ") SS_DELAY(200) SS_TAP(X_ENT));
